Support correlations and searchkeywords in biomarker retrieval
Big refactoring of the biomarker retrieval, it now supports collecting correlated biomarkers and/or search keywords.
The result looks a bit like a poor criteria api on top of HQL, which is unfortunate but the best I could come up with.
The alternative would be building something on top of raw SQL and all the cross-db portability concerns that brings.